+++ src/Products/PloneMeeting/content/meeting.py
@@ -52,6 +52,7 @@ from Products.PloneMeeting.utils import get_annexes
 from Products.PloneMeeting.utils import get_next_meeting
 from Products.PloneMeeting.utils import get_states_before
 from Products.PloneMeeting.utils import getCustomAdapter
+from Products.PloneMeeting.utils import getDateFromDelta
 from Products.PloneMeeting.utils import getWorkflowAdapter
 from Products.PloneMeeting.utils import ItemDuplicatedFromConfigEvent
 from Products.PloneMeeting.utils import MeetingLocalRolesUpdatedEvent
@@ -193,6 +194,16 @@ class IMeeting(IDXMeetingContent):
     form.widget('convocation_date', DatetimeFieldWidget, show_today_link=True, show_time=True)
     convocation_date = schema.Datetime(
         title=_(u'title_convocation_date'),
+        required=False)
+
+    form.widget('validation_deadline', DatetimeFieldWidget, show_today_link=True, show_time=True)
+    validation_deadline = schema.Datetime(
+        title=_(u'title_validation_deadline'),
+        required=False)
+
+    form.widget('freeze_deadline', DatetimeFieldWidget, show_today_link=True, show_time=True)
+    freeze_deadline = schema.Datetime(
+        title=_(u'title_freeze_deadline'),
         required=False)
 
     searchable("place")
@@ -411,7 +422,9 @@ class IMeeting(IDXMeetingContent):
     model.fieldset('dates_and_data',
                    label=_(u"fieldset_dates_and_data"),
                    fields=['date', 'start_date', 'mid_date', 'end_date',
-                           'approval_date', 'convocation_date', 'place', 'place_other',
+                           'approval_date', 'convocation_date',
+                           'validation_deadline', 'freeze_deadline',
+                           'place', 'place_other',
                            'pre_meeting_date', 'pre_meeting_place',
                            'extraordinary_session'])
 
@@ -756,22 +769,30 @@ class Meeting(Container):
     FIELD_INFOS = {
         'date':
             {'optional': False,
-             'condition': ''},
+             'condition': ""},
         'start_date':
             {'optional': True,
-             'condition': ''},
+             'condition': ""},
         'mid_date':
             {'optional': True,
-             'condition': ''},
+             'condition': ""},
         'end_date':
             {'optional': True,
-             'condition': ''},
+             'condition': ""},
         'approval_date':
             {'optional': True,
-             'condition': ''},
+             'condition': ""},
         'convocation_date':
             {'optional': True,
-             'condition': ''},
+             'condition': ""},
+        'validation_deadline':
+            {'optional': True,
+             'condition': "python:context.getTagName() == 'Meeting' and context.date and "
+                "cfg.show_meeting_manager_reserved_field('validation_deadline')"},
+        'freeze_deadline':
+            {'optional': True,
+             'condition': "python:context.getTagName() == 'Meeting' and context.date and "
+                "cfg.show_meeting_manager_reserved_field('freeze_deadline')"},
         'place':
             {'optional': True,
              'condition': ""},
@@ -1785,6 +1806,15 @@ class Meeting(Container):
         # Set, by default, end date to start date + 2 hours.
         if 'end_date' in used_attrs and not self.end_date:
             self.end_date = self.date + timedelta(hours=2)
+        # Compute the deadlines
+        if 'validation_deadline' in used_attrs and not getattr(self, 'validation_deadline', None):
+            delta = cfg.getValidationDeadlineDefault()
+            if not delta.strip() in ('', '0',):
+                self.validation_deadline = getDateFromDelta(self.date, '-' + delta)
+        if 'freeze_deadline' in used_attrs and not getattr(self, 'freeze_deadline', None):
+            delta = cfg.getFreezeDeadlineDefault()
+            if not delta.strip() in ('', '0',):
+                self.freeze_deadline = getDateFromDelta(self.date, '-' + delta)
 
     security.declarePublic('get_user_replacements')