+++ src/Products/PloneMeeting/tests/testMeeting.py
@@ -3846,6 +3846,20 @@ class testMeetingType(PloneMeetingTestCase):
         cfg.setSignatures(three_lines_signatures)
         self.assertFalse(view.warn_assembly(using_attendees=False))
 
+    def test_pm_DeadlineFieldsInit(self):
+        """Test that field Meeting.validation_deadline and Meeting.freeze_deadline
+           are correctly initialized depending on configuration."""
+        cfg = self.meetingConfig
+        self._enableField(('validation_deadline', 'freeze_deadline'), related_to="Meeting")
+        # 5 days before, 9h30
+        self.assertEqual(cfg.getValidationDeadlineDefault(), '5.9:30')
+        # 1 day before, 14h30
+        self.assertEqual(cfg.getFreezeDeadlineDefault(), '1.14:30')
+        self.changeUser("pmManager")
+        meeting = self.create('Meeting', date=datetime(2021, 8, 10))
+        self.assertEqual(meeting.validation_deadline, datetime(2021, 8, 5, 9, 30))
+        self.assertEqual(meeting.freeze_deadline, datetime(2021, 8, 9, 14, 30))
+
 
 def test_suite():
     from unittest import makeSuite